+++ src/inversify.config.ts
@@ -22,6 +22,7 @@ import Disconnect from './commands/disconnect.js';
 import Favorites from './commands/favorites.js';
 import ForwardSeek from './commands/fseek.js';
 import Move from './commands/move.js'
+import NowPlaying from './commands/now-playing.js';
 import Pause from './commands/pause.js';
 import Play from './commands/play.js';
 import QueueCommand from './commands/queue.js';
@@ -67,6 +68,7 @@ container.bind<SpotifyAPI>(TYPES.Services.SpotifyAPI).to(SpotifyAPI).inSingleton
   Favorites,
   ForwardSeek,
   Move,
+  NowPlaying,
   Pause,
   Play,
   QueueCommand,